+++ lib/features/trips/shared/domain/repositories/driver_repository.dart
@@ -1,7 +1,10 @@
 import 'package:delivery_app/features/trips/shared/domain/entities/driver_entity.dart';
+import 'package:delivery_app/features/trips/shared/domain/entities/driver_review_entity.dart';
 
 abstract class DriverRepository {
   Future<List<DriverEntity>> getDrivers();
 
   Future<DriverEntity?> findByName(String name);
+
+  Future<List<DriverReviewEntity>> getReviews(String driverId);
 }